Modify 'initialize' method in container class
Changed the 'initialize' method in the Container class from an instance method to a static one and added the ability to accept a filename as a parameter. Now, instead of always reading a file named 'config.yml', the function can read and parse a YAML file provided by the user.